fix overflow

+++ get_next_line.c
@@ -76,7 +76,7 @@ char	*read_line(int fd, char **lines)
 	char	*buffer;
 	int		byte;
 
-	buffer = ft_calloc(BUFFER_SIZE + 1, sizeof(char));
+	buffer = ft_calloc((size_t) BUFFER_SIZE + 1, sizeof(char));
 	if (!buffer)
 		return (NULL);
 	while (!ft_strchr(*lines, '\n'))
